+++ src/constants/achievements.ts
@@ -1,0 +1,37 @@
+export const ACHIEVEMENTS_DATA = [
+  {
+    id: 1,
+    title: 'First Time Player',
+    description: 'The first time you play the game!',
+  },
+  {
+    id: 2,
+    title: 'Calendar Collector',
+    description: 'Found all the months!',
+  },
+  {
+    id: 3,
+    title: 'Weekly Challenge',
+    description: 'Found all the weekdays!',
+  },
+  {
+    id: 4,
+    title: 'Chemistry Master',
+    description: 'Found all the elements in periodic table!',
+  },
+  {
+    id: 5,
+    title: 'Chemistry Master',
+    description: 'Found all the elements in periodic table!',
+  },
+  {
+    id: 6,
+    title: 'Chemistry Master',
+    description: 'Found all the elements in periodic table!',
+  },
+  {
+    id: 7,
+    title: 'Chemistry Master',
+    description: 'Found all the elements in periodic table!',
+  },
+]