+++ src/main.tsx
@@ -4,12 +4,15 @@ import App from './App.tsx'
 import { ChakraProvider } from '@chakra-ui/react'
 import RecordProvider from './contexts/RecordContext'
 import './index.css'
+import { AchievementProvider } from './contexts/AchievementContext.tsx'
 
 ReactDOM.createRoot(document.getElementById('root')!).render(
   <React.StrictMode>
     <ChakraProvider>
       <RecordProvider>
-        <App />
+        <AchievementProvider>
+          <App />
+        </AchievementProvider>
       </RecordProvider>
     </ChakraProvider>
   </React.StrictMode>,